Sweeping changes for package locations and usage
