+++ src/main/java/io/github/incplusplus/bigtoolbox/network/wlan/AccessPoint.java
@@ -1,6 +1,18 @@
 package io.github.incplusplus.bigtoolbox.network.wlan;
 
+import io.github.incplusplus.bigtoolbox.network.wlan.interop.win.AuthRequestImpl;
+
 public interface AccessPoint {
+	/**
+	 * Document me for the love of god
+	 */
+	boolean connect();
+	
+	boolean connect(AuthRequestImpl authRequest);
+	
+	String getProfileXML();
+	
+	//<editor-fold desc="Getters">
 	String getName();
 	
 	int getSignalStrength();
@@ -16,4 +28,5 @@ public interface AccessPoint {
 	String getCipherAlgorithm();
 	
 	String getBssType();
+	//</editor-fold>
 }
