--- src/main/java/io/github/incplusplus/bigtoolbox/network/wlan/interop/MessageBool.java
@@ -1,23 +1,0 @@
-package io.github.incplusplus.bigtoolbox.network.wlan.interop;
-
-public class MessageBool implements java.io.Serializable
-{
-	public String message;
-	public boolean success;
-
-	public MessageBool(boolean success, String message)
-	{
-		this.success = success;
-		this.message = message;
-	}
-
-	public boolean successful()
-	{
-		return success;
-	}
-
-	public String getMessage()
-	{
-		return message;
-	}
-}
