+++ src/main/java/io/github/incplusplus/bigtoolbox/network/wlan/interop/WLanControllerFactory.java
@@ -1,5 +1,6 @@
 package io.github.incplusplus.bigtoolbox.network.wlan.interop;
 
+import io.github.incplusplus.bigtoolbox.network.wlan.interop.win.WindowsInterop;
 import io.github.incplusplus.bigtoolbox.os.UnsupportedOSException;
 import io.github.incplusplus.bigtoolbox.os.opsys.OperatingSystem;
 
@@ -38,6 +39,8 @@ public final class WLanControllerFactory
 	 * </pre>
 	 * If there already is a non-closed WLanController, close it first before calling this method.
 	 * You should not need to, however, if you use this method within a try-with-resources block as demonstrated.
+	 * You will not be able to use anything related to this {@link WLanController} after the try-with-resources
+	 * block or after calling {@link WLanController#close()} and all objects generated using it will also become unusable.
 	 * @throws IllegalStateException if the factory is in an unexpected state or if
 	 *                         there already is a non-closed WLanController instance.
 	 * @throws UnsupportedOSException if the current OS is not supported by this library
@@ -45,7 +48,6 @@ public final class WLanControllerFactory
 	 */
 	public static WLanController createWLanController()
 	{
-
 		if(numberInstantiated == 0)
 		{
 			if(primaryController != null)
@@ -55,8 +57,8 @@ public final class WLanControllerFactory
 			}
 			else
 			{
-				numberInstantiated++;
 				primaryController=getByOSFamily(OperatingSystem.getOSFamily());
+				numberInstantiated++;
 				return primaryController;
 			}
 		}
@@ -77,12 +79,8 @@ public final class WLanControllerFactory
 				throw new SingletonUnavailableException();
 			}
 		}
-		else
-		{
-			throw new IllegalStateException("The number of instantiated controllers is neither 0 nor 1 " +
-					"despite the fact that this is a singleton factory.");
-		}
-		return null;
+		throw new IllegalStateException("The number of instantiated controllers is neither 0 nor 1 " +
+				"despite the fact that this is a singleton factory.");
 	}
 
 	private static WLanController getByOSFamily(OperatingSystem.OSFamily family)
